Update current account and active account behavior in NAA apps (#7390)
NAA apps do not expect to have multiple accounts, and the account is
supplied by the host application with `accountContext`. This PR allows
for the `accountContext` to reflect in memory and adds checks to make
sure `account` is always mapped with the host's in the NAA app.

This PR also fixes the missing support for forceRefresh and default
cache policy in NAA flows. (merged #7406)

---------

Co-authored-by: baljesingh <baljesingh@microsoft.com>