Rethrow MSAL AuthErrors in NAA flow (#7559)
NAA flow catches errors and runs them through `fromBridgeError`, but not
all errors in the `try` come from the bridge. Notably,
`fromNaaTokenResponse` can throw a `ClientAuthError` if a token is
empty, and cache codepaths can throw some `CacheError`s.

This PR ensures that if an error is already an MSAL `AuthError`, it does
not get passed through `fromBridgeError` again and converted into an
`unknown_error`.